Implement core analysis engine
This adds an analysis engine capable of running analysis programs
composed of analyzer implementations. Analyzers within a program may
depend on one another, and are scheduled in a topological order across a
pool of worker tasks. Dependent analyzers are able to access the output
of their dependencies to adapt their own analysis behavior.